@@ -1,19 +1,16 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import ElementUI from 'element-ui'
-import 'element-ui/lib/theme-chalk/index.css'
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
 import axios from 'axios'
 
-Vue.config.productionTip = false
-
-Vue.use(ElementUI)
+const app = createApp(App)
+app.use(router)
+app.use(ElementPlus)
 
 // 配置axios
 axios.defaults.baseURL = 'http://localhost:8080/api'
-Vue.prototype.$http = axios
+app.config.globalProperties.$http = axios
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app') 
+app.mount('#app') 
